Refactor custom-codec plugin from core repository
This is the initial commit that refactors (copies) the custom-codec
plugin from the core repository into this new downstream opensearch
plugin repository. This is largely a rote copy, with some additional
work to prepare the build.gradle for standalone building. More work
will be needed to onboard the plugin into the official opensearch
bundle.

Signed-off-by: Nicholas Walter Knize <nknize@apache.org>